_settings path only works at index level, not type level
Previously, we were trying to get _settings based on both the index we
were accessing and the type.  _settings is index specific, and not type
specific, so this URL doesn't exist.

This behavior occurred only when using the dump_type action, all other
actions coincidentally had a functional _settings path to fetch.