Support extra metadata fields
This adds support for dumping/restoring the following metadata fields:

* _timestamp
* _version
* _routing
* _percolate
* _parent
* _ttl

Previously these fields weren't included in dumps, and therefore
weren't included in restores.  This was especially problematic for dumps
that included a type with _routing marked as required, causing those
restores to fail.

The format of the new dump files is backwards compatible with the old
restore code (it silently ignores the new fields), and the new restore
is backwards compatible with existing dump files (it only tries to
populate the new metadata if it appears in the dump).