Expose option to dump to preserve version data
The search api expects `version: true` to be set in the params, not as a
requested field.

The bulk api expects both `"_version":<version_number>` and `"version_type": "external"` if
something other than automatic versions are being used.

Since the underlying es apis don't have version preserving as the default option, this
exposes a new `--preserve-versions` option to the dump and dump_type
subcommands, with a default of false

Doc references:
https://www.elastic.co/guide/en/elasticsearch/reference/6.0/search-request-version.html
https://www.elastic.co/guide/en/elasticsearch/reference/2.4/docs-bulk.html#bulk-versioning

Fixes #25